Add clearance configuration and fix bugs
Add configuration file and clearance precondition for validating that a user has permission to use the bot.
Load the prefix from the configuration instead of hard coding it.
Fix naming collision between the Copy command and CommandSet by renaming the command to CopyTarget.
Apply the loaded embed precondition to the ClearGroup CommandSet, as this was forgotten when the commands were moved.