Misc changes
Allow the Clear command argument to be optional, so that providing no arguments prints the options.
Add file sending to the export command to allow exporting large embeds. Format short JSON with a code block.
Correct the response for a failed CloneCluster operation.
Remove redundant requiresGuild field in all commands. This is covered by preconditions.